+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Support for HBase 2.4.9 ([#133]).
 - Support for HBase 2.4.11 ([#148]).
+- Support for HBase 2.4.12 ([#197]).
 - Use cli argument `watch-namespace` / env var `WATCH_NAMESPACE` to specify
   a single namespace to watch ([#137]).
 - Writing a discovery config map containing `hbase-site.xml` with the `hbase.zookeeper.quorum` property ([#163]).
@@ -26,6 +27,7 @@
 [#162]: https://github.com/stackabletech/hbase-operator/pull/162
 [#163]: https://github.com/stackabletech/hbase-operator/pull/163
 [#179]: https://github.com/stackabletech/hbase-operator/pull/179
+[#197]: https://github.com/stackabletech/hbase-operator/pull/197
 
 ## [0.2.0] - 2022-02-14