+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## Changed
+
+- Include chart name when installing with a custom release name ([#209], [#210]).
+
+[#209]: https://github.com/stackabletech/hbase-operator/pull/209
+[#210]: https://github.com/stackabletech/hbase-operator/pull/210
+
 ## [0.3.0] - 2022-06-30
 
 ### Added