+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ## Changed
 
 - Startup probe created and thresholds in liveness and readiness probes fine-tuned ([#193]).
-- Orphaned resources are deleted
+- Orphaned resources are deleted ([#215]).
 
 [#193]: https://github.com/stackabletech/hbase-operator/pull/193
+[#215]: https://github.com/stackabletech/hbase-operator/pull/215
 
 ## [0.3.0] - 2022-06-30