+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 ## Changed
 
 - Startup probe created and thresholds in liveness and readiness probes fine-tuned ([#193]).
+- Include chart name when installing with a custom release name ([#209], [#210]).
 - Orphaned resources are deleted ([#215]).
 
 [#193]: https://github.com/stackabletech/hbase-operator/pull/193
+[#209]: https://github.com/stackabletech/hbase-operator/pull/209
+[#210]: https://github.com/stackabletech/hbase-operator/pull/210
 [#215]: https://github.com/stackabletech/hbase-operator/pull/215
 
 ## [0.3.0] - 2022-06-30