+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 - Startup probe created and thresholds in liveness and readiness probes fine-tuned ([#193]).
 - Include chart name when installing with a custom release name ([#209], [#210]).
 - Orphaned resources are deleted ([#215]).
+- Fix HBase-shell start failure ([#218]).
 
 [#193]: https://github.com/stackabletech/hbase-operator/pull/193
 [#209]: https://github.com/stackabletech/hbase-operator/pull/209
 [#210]: https://github.com/stackabletech/hbase-operator/pull/210
 [#215]: https://github.com/stackabletech/hbase-operator/pull/215
+[#218]: https://github.com/stackabletech/hbase-operator/pull/218
 
 ## [0.3.0] - 2022-06-30