+++ CHANGELOG.md
@@ -4,12 +4,16 @@
 
 ## Changed
 
+- Startup probe created and thresholds in liveness and readiness probes fine-tuned ([#193]).
 - Include chart name when installing with a custom release name ([#209], [#210]).
+- Orphaned resources are deleted ([#215]).
 - Fix HBase-shell start failure ([#218]).
 - Add integration tests and usage documentation for Phoenix ([#221]).
 
+[#193]: https://github.com/stackabletech/hbase-operator/pull/193
 [#209]: https://github.com/stackabletech/hbase-operator/pull/209
 [#210]: https://github.com/stackabletech/hbase-operator/pull/210
+[#215]: https://github.com/stackabletech/hbase-operator/pull/215
 [#218]: https://github.com/stackabletech/hbase-operator/pull/218
 [#221]: https://github.com/stackabletech/hbase-operator/pull/221
 
@@ -32,7 +36,6 @@
   - `hbaseManagesZk`: defaults to false
   - `hbaseClusterDistributed`: defaults to true
 - [BREAKING] Specifying the product version has been changed to adhere to [ADR018](https://docs.stackable.tech/home/contributor/adr/ADR018-product_image_versioning.html) instead of just specifying the product version you will now have to add the Stackable image version as well, so `version: 3.5.8` becomes (for example) `version: 3.5.8-stackable0.1.0` ([#179])
-- Startup probe created and thresholds in liveness and readiness probes fine-tuned ([#193]).
 
 [#133]: https://github.com/stackabletech/hbase-operator/pull/133
 [#137]: https://github.com/stackabletech/hbase-operator/pull/137
@@ -41,7 +44,6 @@
 [#162]: https://github.com/stackabletech/hbase-operator/pull/162
 [#163]: https://github.com/stackabletech/hbase-operator/pull/163
 [#179]: https://github.com/stackabletech/hbase-operator/pull/179
-[#193]: https://github.com/stackabletech/hbase-operator/pull/193
 [#197]: https://github.com/stackabletech/hbase-operator/pull/197
 
 ## [0.2.0] - 2022-02-14